Created a basic node app with some routing
The node app is created by calling express(). A router folder was created to hold the routes for each feature. A routes file for Products was created ad is currently holding two routes: for all products, and specific products. Every response back to the client will be json. The app.js file is using the productsRoutes as middleware.