Added User Schema. Implemented Cart routes
Created a User model based on userSchema. The userSchema defines a relationship in the cart.items path between User and Product. Added two new routes for cart: Adding product to user cart, and getting user cart. The addToCart method, is a method on the userSchema, which adds a product to the cart if not already existing, otherwise it adds in quantity. Added addToCart and getCart route handler functions in the products-controller.