Created Order model, routes, and controller functions
Created an Order model which will return a user and products/items in their cart. The routes and functions are located in the products routes and controller, but will be moved in the future. Dummy user is being use in many functions, and will need to be replaced in the future.